+++ rust/lancedb/src/remote/client.rs
@@ -4,7 +4,7 @@
 use std::{collections::HashMap, future::Future, str::FromStr, time::Duration};
 
 use http::HeaderName;
-use log::debug;
+use log::{debug, info};
 use reqwest::{
     header::{HeaderMap, HeaderValue},
     Request, RequestBuilder, Response,
@@ -12,6 +12,7 @@ use reqwest::{
 
 use crate::error::{Error, Result};
 use crate::remote::db::RemoteOptions;
+use crate::remote::retry::ResolvedRetryConfig;
 
 const REQUEST_ID_HEADER: HeaderName = HeaderName::from_static("x-request-id");
 
@@ -124,36 +125,6 @@ pub struct RetryConfig {
     /// The default is 409, 429, 500, 502, 503, 504.
     pub statuses: Option<Vec<u16>>,
     // TODO: should we allow customizing methods?
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct ResolvedRetryConfig {
-    pub(crate) retries: u8,
-    pub(crate) connect_retries: u8,
-    pub(crate) read_retries: u8,
-    pub(crate) backoff_factor: f32,
-    pub(crate) backoff_jitter: f32,
-    pub(crate) statuses: Vec<reqwest::StatusCode>,
-}
-
-impl TryFrom<RetryConfig> for ResolvedRetryConfig {
-    type Error = Error;
-
-    fn try_from(retry_config: RetryConfig) -> Result<Self> {
-        Ok(Self {
-            retries: retry_config.retries.unwrap_or(3),
-            connect_retries: retry_config.connect_retries.unwrap_or(3),
-            read_retries: retry_config.read_retries.unwrap_or(3),
-            backoff_factor: retry_config.backoff_factor.unwrap_or(0.25),
-            backoff_jitter: retry_config.backoff_jitter.unwrap_or(0.25),
-            statuses: retry_config
-                .statuses
-                .unwrap_or_else(|| vec![409, 429, 500, 502, 503, 504])
-                .into_iter()
-                .map(|status| reqwest::StatusCode::from_u16(status).unwrap())
-                .collect(),
-        })
-    }
 }
 
 // We use the `HttpSend` trait to abstract over the `reqwest::Client` so that
@@ -413,7 +384,7 @@ impl<S: HttpSend> RestfulLanceDbClient<S> {
         Ok((request_id, response))
     }
 
-    /// extract the request ID from the request headers.
+    /// Extract the request ID from the request headers.
     /// If the request ID header is not set, this will generate a new one and set
     /// it on the request headers
     pub fn extract_request_id(&self, request: &mut Request) -> String {
@@ -423,80 +394,19 @@ impl<S: HttpSend> RestfulLanceDbClient<S> {
             request_id.to_str().unwrap().to_string()
         } else {
             let request_id = uuid::Uuid::new_v4().to_string();
-            let header = HeaderValue::from_str(&request_id).unwrap();
-            request.headers_mut().insert(REQUEST_ID_HEADER, header);
+            self.set_request_id(request, &request_id);
             request_id
         };
+
+        info!("{:#?}", request.headers());
         request_id
     }
-    //
-    // async fn send_with_retry( // todo: remove
-    //     &self,
-    //     client: reqwest::Client,
-    //     req: Request,
-    //     request_id: String,
-    //     retry_5xx: bool,
-    // ) -> Result<(String, Response)> {
-    //     let mut retry_counter = RetryCounter::new(&self.retry_config, request_id);
-    //
-    //     let non_5xx_statuses = self
-    //         .retry_config
-    //         .statuses
-    //         .iter()
-    //         .filter(|s| !s.is_server_error())
-    //         .cloned()
-    //         .collect::<Vec<_>>();
-    //     loop {
-    //         // This only works if the request body is not a stream. If it is
-    //         // a stream, we can't use the retry path. We would need to implement
-    //         // an outer retry.
-    //         let request = req.try_clone().ok_or_else(|| Error::Runtime {
-    //             message: "Attempted to retry a request that cannot be cloned".to_string(),
-    //         })?;
-    //         let response = self
-    //             .sender
-    //             .send(&client, request)
-    //             .await
-    //             .map(|r| (r.status(), r));
-    //         match response {
-    //             Ok((status, response)) if status.is_success() => {
-    //                 debug!(
-    //                     "Received response for request_id={}: {:?}",
-    //                     retry_counter.request_id, &response
-    //                 );
-    //                 return Ok((retry_counter.request_id, response));
-    //             }
-    //             Ok((status, response))
-    //                 if (retry_5xx && self.retry_config.statuses.contains(&status))
-    //                     || non_5xx_statuses.contains(&status) =>
-    //             {
-    //                 let source = self
-    //                     .check_response(&retry_counter.request_id, response)
-    //                     .await
-    //                     .unwrap_err();
-    //                 retry_counter.increment_request_failures(source)?;
-    //             }
-    //             Err(err) if err.is_connect() => {
-    //                 retry_counter.increment_connect_failures(err)?;
-    //             }
-    //             Err(err) if err.is_timeout() || err.is_body() || err.is_decode() => {
-    //                 retry_counter.increment_read_failures(err)?;
-    //             }
-    //             Err(err) => {
-    //                 let status_code = err.status();
-    //                 return Err(Error::Http {
-    //                     source: Box::new(err),
-    //                     request_id: retry_counter.request_id,
-    //                     status_code,
-    //                 });
-    //             }
-    //             Ok((_, response)) => return Ok((retry_counter.request_id, response)),
-    //         }
-    //
-    //         let sleep_time = retry_counter.next_sleep_time();
-    //         tokio::time::sleep(sleep_time).await;
-    //     }
-    // }
+
+    /// Set the request ID header
+    pub fn set_request_id(&self, request: &mut Request, request_id: &str) {
+        let header = HeaderValue::from_str(&request_id).unwrap();
+        request.headers_mut().insert(REQUEST_ID_HEADER, header);
+    }
 
     pub async fn check_response(&self, request_id: &str, response: Response) -> Result<Response> {
         // Try to get the response text, but if that fails, just return the status code
@@ -516,91 +426,6 @@ impl<S: HttpSend> RestfulLanceDbClient<S> {
                 status_code: Some(status),
             })
         }
-    }
-}
-
-pub struct RetryCounter<'a> {
-    pub request_failures: u8,
-    pub connect_failures: u8,
-    pub read_failures: u8,
-    pub config: &'a ResolvedRetryConfig,
-    pub request_id: String,
-}
-
-impl<'a> RetryCounter<'a> {
-    fn new(config: &'a ResolvedRetryConfig, request_id: String) -> Self {
-        Self {
-            request_failures: 0,
-            connect_failures: 0,
-            read_failures: 0,
-            config,
-            request_id,
-        }
-    }
-
-    fn check_out_of_retries(
-        &self,
-        source: Box<dyn std::error::Error + Send + Sync>,
-        status_code: Option<reqwest::StatusCode>,
-    ) -> Result<()> {
-        if self.request_failures >= self.config.retries
-            || self.connect_failures >= self.config.connect_retries
-            || self.read_failures >= self.config.read_retries
-        {
-            Err(Error::Retry {
-                request_id: self.request_id.clone(),
-                request_failures: self.request_failures,
-                max_request_failures: self.config.retries,
-                connect_failures: self.connect_failures,
-                max_connect_failures: self.config.connect_retries,
-                read_failures: self.read_failures,
-                max_read_failures: self.config.read_retries,
-                source,
-                status_code,
-            })
-        } else {
-            Ok(())
-        }
-    }
-
-    pub fn increment_request_failures(&mut self, source: crate::Error) -> Result<()> {
-        self.request_failures += 1;
-        let status_code = if let crate::Error::Http { status_code, .. } = &source {
-            *status_code
-        } else {
-            None
-        };
-        self.check_out_of_retries(Box::new(source), status_code)
-    }
-
-    pub fn increment_connect_failures(&mut self, source: reqwest::Error) -> Result<()> {
-        self.connect_failures += 1;
-        let status_code = source.status();
-        self.check_out_of_retries(Box::new(source), status_code)
-    }
-
-    pub fn increment_read_failures(&mut self, source: reqwest::Error) -> Result<()> {
-        self.read_failures += 1;
-        let status_code = source.status();
-        self.check_out_of_retries(Box::new(source), status_code)
-    }
-
-    pub fn next_sleep_time(&self) -> Duration {
-        let backoff = self.config.backoff_factor * (2.0f32.powi(self.request_failures as i32));
-        let jitter = rand::random::<f32>() * self.config.backoff_jitter;
-        let sleep_time = Duration::from_secs_f32(backoff + jitter);
-        debug!(
-            "Retrying request {:?} ({}/{} connect, {}/{} read, {}/{} read) in {:?}",
-            self.request_id,
-            self.connect_failures,
-            self.config.connect_retries,
-            self.request_failures,
-            self.config.retries,
-            self.read_failures,
-            self.config.read_retries,
-            sleep_time
-        );
-        sleep_time
     }
 }
